Refactor config helpers for better delete and create

+++ drivy_tools/src/utils/config_utils.py
@@ -8,10 +8,11 @@ from drivy_tools.state import state
 
 
 def create_drivy_folder_and_config_file(config: Config):
-    if not default_drivy_tools_dir.exists():
+    config_file_dir = default_drivy_tools_dir.joinpath("config").with_suffix(".ini")
+    if not config_file_dir.exists():
         is_config_wanted = typer.confirm("Config file doesn't exist! Do you want to create?")
         if is_config_wanted:
-            default_drivy_tools_dir.mkdir()
+            default_drivy_tools_dir.mkdir(exist_ok=True)
             config_ = configparser.ConfigParser()
             edit = typer.confirm("Do you want to change default values?")
             for key, value in config.dict().items():
@@ -28,15 +29,17 @@ def create_drivy_folder_and_config_file(config: Config):
         return None
 
 
-def delete_config_folder():
+def delete_config_folder(delete_folder: bool = False):
     for thing in default_drivy_tools_dir.iterdir():
         if thing.is_file():
             thing.unlink()
         else:
-            for th in thing.iterdir():
-                th.unlink()
-            thing.rmdir()
-    default_drivy_tools_dir.rmdir()
+            if delete_folder:
+                for th in thing.iterdir():
+                    th.unlink()
+                thing.rmdir()
+    if delete_folder:
+        default_drivy_tools_dir.rmdir()
     if state.verbose:
         print("Config file is deleted!")
     return None
